PR template (#5)
* feat: Add PR template
