+++ .github/pull_request_template.md
@@ -1,0 +1,30 @@
+### Fixes #<ISSUE_NUMBER>
+
+### Description
+Include a summary of the change and relevant motivation/context. List any dependencies that are required for this change.
+
+
+**Type of change** 
+
+- [ ] New feature (non-breaking change which adds functionality pre-approved by mentors)
+- [ ] Bug fix (non-breaking change which fixes an issue)
+- [ ] Refactor
+- [ ] Addition of testcases
+- [ ] This change requires a documentation update (software upgrade on readme file)
+
+
+### Screenshots of change:
+Describe the tests you ran to verify your changes. Provide instructions or GIFs so we can reproduce. List any relevant details for your test.
+
+
+### Checklist:
+- [ ] My PR follows the style guidelines of this project
+- [ ] I have commented my code or provided relevant documentation, particularly in hard-to-understand areas
+- [ ] I have updated storybook
+- [ ] I have updated README.md
+- [ ] Added new npm packages
+
+**Code/Quality Assurance**
+- [ ] My changes generate no new warnings 
+- [ ] My PR currently breaks something (fix or feature that would cause existing functionality to not work as expected)
+- [ ] New and existing tests pass locally with my changes
