+++ .storybook/main.js
@@ -1,0 +1,12 @@
+module.exports = {
+  stories: ['../src/**/*.stories.js'],
+  addons: [
+    '@storybook/preset-create-react-app',
+    '@storybook/addon-actions',
+    '@storybook/addon-links',
+    '@storybook/addon-storysource',
+    '@storybook/addon-docs',
+    '@storybook/addon-knobs',
+    '@storybook/addon-viewport',
+  ],
+};
