+++ src/index.js
@@ -5,7 +5,7 @@ import App from './App';
 import * as serviceWorker from './serviceWorker';
 
 import {Provider} from 'react-redux';
-import {store} from './store/createStore';
+import {store} from './createStore';
 
 ReactDOM.render(
   <React.StrictMode>
