+++ src/reducers/RootReducer.js
@@ -1,5 +1,5 @@
 import {combineReducers} from 'redux';
-import success from './successReducer';
+import success from './SuccessReducer';
 
 export default combineReducers({
     success
